requirements.txt
streamlit==1.32.0
python-pptx==0.6.21
Pillow==10.2.0
requests==2.31.0
python-magic==0.4.27
python-magic-bin==0.4.14; platform_system == "Windows"
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.4
ollama==0.1.2
pillow==9.5.0
matplotlib==3.7.1
numpy==1.24.3
transformers==4.28.1
torch>=2.1.0
wcag-contrast-ratio==0.9
cairosvg==2.7.0 